Notify HTTPHandler instances of errors (#20)
* Notify HTTPHandler instances of errors

and cancellation of requests.

This includes failed auth (e.g. ssl), failed to connect, prematurely closed connection.

* add keepalive_timeout_secs parameter to client

* Add timeout parameter to httpget example

* WIP test error handling

* Finish test for error handling for HTTP client

* add server error handling test

for connection-closed failures only as all others cannot be reported for the server side.

* avoid creating a plaintext session when ssl session initiation fails

and improve performance by avoiding a 'casting'-error as control flow where the
error case is the normal path in most cases.