Add regex and net_ssl as dependencies
This commit addresses the fact that regex and net/ssl were removed
from the Pony standard library. Since the http library depends on
these libraries we need to include them or the tests and examples
won't build.

The dependencies are added to a `bundle.json` file in the top level of
the project. This makes it available to all of the build targets
(test, bench, and examples).

Closes #23