Add additional test environments
Prior to this commit, we were only testing against a single SSL library
on Linux plus also a version on Windows.

We have seen that with http related code, we can encounter errors on
different SSL versions. See https://github.com/ponylang/net_ssl/issues/105.
To adjust to this, with this test, we are adding testing against all our
main supported SSL versions on Linux.

This does not change our MacOS (not currently done) or our Windows testing
matrix for ponylang/http.