Change handler factory creation of HTTPClient (issue #102)
The handler factory was supplied in the apply method of a client
together with the Payload for the request. If the client
already had a session for the new host, port and scheme, no new
session was created and the supplied handler factory wasn't used.

The new design makes it clear, that one client uses the same handler
factory for all requests it makes. If different handler factories are
needed, different clients need to be created.

Closes #102
Closes #9